refactor: Move alerts-grouping logic to the backend
